DCCTEST-2547 Updated icgc_relations.tsv to match runme.pl
All of the relations in the TSV file were rearranged to be in the same order as the joins in runme.pl. "First" and "Second" table/key were swapped to match their ordering in runme.pl. For the gene table joins (at the end of the file), the key names were changed according to Junjun and Anthony's instructions. A total of four relations were removed: three of them dealt with germline secondary files, which do not exist, and the last was a join between germline and samples, which does not occur.